refactor: Remove Local Login Redundancies
